8 şubat 2024

+++ Entities/DTO/CarDetailDto.cs
@@ -18,6 +18,6 @@ namespace Entities.DTO
 
         public int ModelYear { get; set; }
         public string Description { get; set; }
-        public double DailyPrice { get; set; }
+        public int DailyPrice { get; set; }
     }
 }
